add new tasks

+++ src/LinguaLeo/Tools/wti/Tasks/SaveWtiCommentsTask.php
@@ -4,18 +4,12 @@ namespace LinguaLeo\Tools\wti\Tasks;
 
 require __DIR__ . '/../../../../../vendor/autoload.php';
 
-use LinguaLeo\Tools\Task;
-use LinguaLeo\wti\WtiApi;
+use LinguaLeo\Tools\WtiTask;
 
-class SaveWtiCommentsTask extends Task
+class SaveWtiCommentsTask extends WtiTask
 {
 
     protected $outputFolder;
-
-    /**
-     * @var WtiApi
-     */
-    protected $wti;
 
     protected function init()
     {
@@ -30,8 +24,14 @@ class SaveWtiCommentsTask extends Task
         $this->saveComments();
     }
 
+    protected function getFolderName() {
+        return $this->outputFolder . $this->wti->getProjectInfo()->id;
+    }
+
     protected function saveComments() {
-        $this->deleteDir($this->getFolderName());
+        if (is_dir($this->getFolderName())) {
+            $this->deleteDir($this->getFolderName());
+        }
 
         foreach ($this->wti->getProjectInfo()->project_files as $projectFile) {
             $comments = array();
@@ -59,12 +59,6 @@ class SaveWtiCommentsTask extends Task
         }
     }
 
-    protected function prepareFolder($folder) {
-        if (!is_dir($folder)) {
-            mkdir($folder, 0777, true);
-        }
-    }
-
     protected function saveCommentsFile($comments, $file) {
         if (!count($comments)) {
             return;
@@ -79,50 +73,6 @@ class SaveWtiCommentsTask extends Task
         $this->prepareFolder(pathinfo($filePath, PATHINFO_DIRNAME));
 
         file_put_contents($filePath, json_encode($comments));
-    }
-
-    protected function getFolderName() {
-        return $this->outputFolder . $this->wti->getProjectInfo()->id;
-    }
-
-    /**
-     * @param $apiKey
-     * @param bool $initProjectInfo
-     * @return WtiApi
-     */
-    protected function getWti($apiKey, $initProjectInfo = true)
-    {
-        return new WtiApi($apiKey, $initProjectInfo);
-    }
-
-    private function getPassedApiKey()
-    {
-        try {
-            $apiKey = $this->getArgument(0);
-        } catch (\InvalidArgumentException $exception) {
-            $this->error('Please provide an API key!');
-            exit;
-        }
-        return $apiKey;
-    }
-
-
-    public static function deleteDir($dirPath) {
-        if (! is_dir($dirPath)) {
-            throw new InvalidArgumentException("$dirPath must be a directory");
-        }
-        if (substr($dirPath, strlen($dirPath) - 1, 1) != '/') {
-            $dirPath .= '/';
-        }
-        $files = glob($dirPath . '*', GLOB_MARK);
-        foreach ($files as $file) {
-            if (is_dir($file)) {
-                self::deleteDir($file);
-            } else {
-                unlink($file);
-            }
-        }
-        rmdir($dirPath);
     }
 }
